+++ deeplake/integrations/mmlab/mmseg/registry.py
@@ -1,0 +1,8 @@
+from mmseg.registry import TRANSFORMS as MMSEG_TRANSFORMS
+from mmengine.registry import Registry
+
+TRANSFORMS = Registry(
+    "transform",
+    parent=MMSEG_TRANSFORMS,
+    locations=["deeplake.integrations.mmlab.mmseg.mmseg_"],
+)
